requirements.txt
customtkinter
Pillow



(pip install ...)